Link to helpful instructions for using Docker for Mac's Kubernetes support for testing
I wrote these instructions for myself for testing our YAML here on a Docker for Mac Kubernetes cluster. I am not sure if these are beneficial for others on the team or not, so I kept the section very small and linked to a separate gist to really drive home that these are more personal notes.

In fact, I'm not sure if merging this is worth it. Thoughts? At the very least, I am opening this PR so it is findable via GitHub PR search.