Add ingress-nginx and an ingress for frontend
For 3.0 we have removed many of our HTTP and TLS features from the frontend
binary. Instead we have moved those responsibilities to nginx. This should
give more power to end users to configure Sourcegraph to suit there internal
needs, without needing to wait for us to ship product features.

This commit adds ingress-nginx to the configure environment which users should
use when setting up the cluster. We do not add it to base since it may
conflict with existing ingress-nginx already present in the cluster, as well
as requiring some configuration depending on cloud provider.

Additionally we include an ingress for Sourcegraph frontend.