Add ingress-nginx and an ingress for frontend (#159)
* Add ingress-nginx and an ingress for frontend

For 3.0 we have removed many of our HTTP and TLS features from the frontend
binary. Instead we have moved those responsibilities to nginx. This should
give more power to end users to configure Sourcegraph to suit there internal
needs, without needing to wait for us to ship product features.

This commit adds ingress-nginx to the configure environment which users should
use when setting up the cluster. We do not add it to base since it may
conflict with existing ingress-nginx already present in the cluster, as well
as requiring some configuration depending on cloud provider.

Additionally we include an ingress for Sourcegraph frontend.

* Apply suggestions from code review

Co-Authored-By: keegancsmith <keegan.csmith@gmail.com>

* nginx: Remove https (30081 node port) from frontend service

We no longer expose HTTPS directly on frontend.

* nginx: Remove proxy-{send,read}-timeout adjustment

We don't have any websocket connections to the frontend. We only have them to
extensions, so we can remove the adjustment to the timeouts. (Default is 60s).