indexed-search: Migrate to a StatefulSet
Includes documentation on how to retain the existing volume.

Note: We do not yet give information on how to scale indexed-search. This will
come once horizontal sharding support is in master. Until then a statefulset
of size 1 will act in the same way as the previous deployment with size 1.

Test Plan: Tested on minikube via upgrading an older cluster.