frontend: Do not specify ZOEKT_HOST
The 3.10 docker images of Sourcegraph use service discovery, so do not require
ZOEKT_HOST to be specified. Additionally, the default has been updated to what
the value was specified as. This means we do not need to include this value.

Adding ZOEKT_HOST was added by
https://github.com/sourcegraph/deploy-sourcegraph/pull/428 but was made
obsolete by https://github.com/sourcegraph/sourcegraph/pull/6296. It being
specified just means we do not use service discovery, and does not lead to any
oter problems.