search: generous timeout for search pods
These services when searching will use a lot of CPU and IO. This may
make them slow to respond. The readiness probe failing has disasterous
consequences, since k8s' service discovery will remove the pod from the
endpoint list leading to us rebalancing.

This is a guess at a root cause of problems we are seeing on a very
large instance.