add codeinsights-db (TimescaleDB) deployment
This PR adds the Code Insights TimescaleDB deployment, which will be used to
store time series data for tracking insights about code. For more information
on what this will be used for, see https://github.com/sourcegraph/sourcegraph/issues/17218

TimescaleDB is a plugin for Postgres which extends it to support time series
data storage. We only use the OSS version, and so deploying it is always free
(and their non-free version supports our use case anyway.)

People deploying Sourcegraph should use this deployment and not try to deploy
TimescaleDB as a plugin in their Postgres deployment for a few reasons:

1. TimescaleDB only supports specific Postgres versions, e.g. not Postgres 13
   yet.
2. TimescaleDB should have isolated resources from the rest of Sourcegraph, so
   as to prevent Code Insights from interfering with other Sourcegraph features
   in general.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>