frontend: Remove SRC_GIT_SERVERS env var (#3744)
In Kubernetes deployments, setting the SRC_GIT_SERVERS environment
variable explicitly is no longer needed. Addresses of the gitserver pods
will be discovered automatically and in the same numerical order as with
the static list. Unset the env var in your frontend.Deployment.yaml
to make use of this feature. sourcegraph/sourcegraph#24094