ci: output which file is failing verify-labels
A file was failing to parse with yj. This was harder than it should of
been to track down. This commit improves the verify-label script to run
per file so we know which file is responsible.

Other improvements:
- Check if the value is sourcegraph, not containing sourcegraph. Uses jq
  equality check.
- Don't install yj if already present. Useful for testing locally.
- Use bash on env. bash at /bin can be old (and doesn't exist on NixOS).

Test Plan: modified some yaml files to fail the rule and ran
.buildkite/verify-label.sh.